// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const task_status = {
  not_started: 'not_started',
  in_progress: 'in_progress',
  completed: 'completed',
  cancelled: 'cancelled'
} as const

export type task_status = (typeof task_status)[keyof typeof task_status]


export const provider_type = {
  local: 'local',
  facebook: 'facebook',
  googlefb_and_google: 'googlefb_and_google'
} as const

export type provider_type = (typeof provider_type)[keyof typeof provider_type]


export const subtask_status = {
  done: 'done',
  not_done: 'not_done'
} as const

export type subtask_status = (typeof subtask_status)[keyof typeof subtask_status]


export const task_priority = {
  low: 'low',
  high: 'high',
  critical: 'critical'
} as const

export type task_priority = (typeof task_priority)[keyof typeof task_priority]